requirements.txt
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.4
diskcache>=5.6.0
python-dotenv>=1.0.0
google-genai>=0.3.0